Potential fix to multi-shot spawn
After the first bullet is set on the board, the coroutine doing the waiting is called dozens of times since it is in the update function. The fix suggested by this commit is to add a flag the limits the amount on coroutine instances to 1, instead of doezens. Theoretically, this should spawn only one shot now.